Re arrange the class structure with new java files
