+++ bin/.gitignore
@@ -2,3 +2,6 @@
 /LineFollowing.class
 /LineFollowing$ColorDetectionThread.class
 /LineFollowing$LineFollowingThread.class
+/ColorDetectionThread.class
+/ObstacleDetectionThread.class
+/LineFollowingThread.class
